+++ tests/unit/model/test_model_item.py
@@ -45,6 +45,25 @@ def test_model_item_init(attributes):
         assert getattr(model_item, attr) == expected
 
 
+def test_handling_of_bogus_init_params():
+    """
+    Test for sensible error message if wrong param passed.
+
+    See https://github.com/Midnighter/structurizr-python/issues/50.
+    """
+    with pytest.raises(
+        TypeError, match=r"ConcreteModelItem.__init__\(\) got an unexpected "
+                         r"keyword argument 'foo'"
+    ):
+        ConcreteModelItem(foo=7)
+    with pytest.raises(
+        TypeError,
+        match=r"ConcreteModelItem.__init__\(\) got unexpected keyword "
+              r"arguments 'foo', 'bar'",
+    ):
+        ConcreteModelItem(foo=7, bar=17)
+
+
 def test_default_element_tags_order(empty_model: Model):
     """
     Test that the default tags get added in the right order.